# frontend/openapi (1).json
{"openapi":"3.1.0","info":{"title":"Medical Imaging Noise Reduction Virtual Lab API","description":"Backend API for running denoising autoencoder experiments.","version":"0.1.0"},"paths":{"/api/dataset/preview":{"get":{"tags":["Experiment"],"summary":"Preview Dataset","operationId":"preview_dataset_api_dataset_preview_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string","default":"mnist_784","title":"Name"}},{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":16,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Experiment"],"summary":"Health Check","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"anyOf":[{"type":"boolean"},{"type":"string"}]},"type":"object","title":"Response Health Check  Get"}}}}}}},"/models":{"get":{"tags":["Experiment"],"summary":"Get Models","operationId":"get_models_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response Get Models Models Get"}}}}}}},"/metrics":{"get":{"tags":["Experiment"],"summary":"Get Metrics","operationId":"get_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/sample-images":{"get":{"tags":["Experiment"],"summary":"Get Sample Images","operationId":"get_sample_images_sample_images_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Get Sample Images Sample Images Get"}}}}}}},"/run-experiment":{"post":{"tags":["Experiment"],"summary":"Run Experiment","operationId":"run_experiment_run_experiment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunExperimentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RunExperimentRequest":{"properties":{"model_type":{"type":"string","const":"denoising_autoencoder","title":"Model Type"},"noise_factor":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Noise Factor","default":0.3},"latent_dim":{"type":"integer","maximum":512.0,"minimum":2.0,"title":"Latent Dim","default":64}},"type":"object","required":["model_type"],"title":"RunExperimentRequest","description":"Input payload for experiment simulation endpoint."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}